Apply Sequence/BinkMovieSys header improvements from WIP (+336 bytes, +4 functions)
Added Sequence.h, Sequence_p.h, BinkMovieSys.h header improvements.
Milo Engine: 3,250,372 → 3,250,708 bytes, 19,847 → 19,851 functions.

+++ src/system/synth/Sequence_p.h
@@ -71,8 +71,8 @@ public:
 
 protected:
     int unk40;
-    int unk44;
-    int unk48;
+    float unk44;
+    float unk48;
     std::vector<float> unk4c;
     bool unk54;
 };
